fix(crt_refptrs): read string table offset from Name.Long, not Name.Short
For COFF symbols with names > 8 chars, the string table offset lives in
bytes 4-7 of the name field (Name.Long), not bytes 0-3 (Name.Short).
This matches the existing implementation in get_symbol_name() in pe_parser.c.

+++ src/stubs/crt_refptrs.c
@@ -242,7 +242,7 @@ static uint64_t find_symbol_rva_from_file(const char *file_path,
             sym_name = (const char *)sym->N.ShortName;
             sym_name_len = strlen(sym_name);
         } else if (string_table) {
-            uint32_t offset = sym->N.Name.Short;
+            uint32_t offset = sym->N.Name.Long;
             if (offset > 0) {
                 sym_name = string_table + offset;
                 sym_name_len = strlen(sym_name);
